requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
SQLAlchemy==2.0.36
mysqlclient==2.2.5
aiomysql==0.2.0
redis==5.1.1
passlib[bcrypt]==1.7.4
python-jose==3.3.0
pydantic==2.9.2
pydantic-settings==2.6.1
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
itsdangerous==2.2.0
python-multipart==0.0.12
bcrypt==4.1.2
alembic==1.17.1
email-validator==2.3.0
python-dateutil>=2.8.2